--- ReportHub/ReportHub.Infrastructure/Services/FileServices/CsvBaseService.cs
@@ -1,35 +1,0 @@
-﻿using CsvHelper;
-using System.Reflection;
-
-namespace ReportHub.Infrastructure.Services.FileServices;
-
-public abstract class CsvBaseService
-{
-    /// <summary>
-    /// Get count of all properties that exist in csv file
-    /// </summary>
-    /// <param name="reader"></param>
-    /// <param name="properties"></param>
-    /// <returns></returns>
-    private protected static int GetAllExistPropertyInCsv(CsvReader reader, PropertyInfo[] properties)
-    {
-        var count = 0;
-        for (var i = 0; i < properties.Length; i++)
-        {
-            var prop = properties[i];
-            var isField = reader.TryGetField(prop.PropertyType, prop.Name, out object field);
-            if (!isField)
-                break;
-            count++;
-        }
-
-        return count;
-    }
-
-    /// <summary>
-    /// Get properties of generic type
-    /// </summary>
-    /// <typeparam name="T"></typeparam>
-    /// <returns></returns>
-    private protected PropertyInfo[] GetTypeProperties<T>() => typeof(T).GetProperties();
-}
